src/crawler/requirements_docker.txt
beautifulsoup4==4.13.3
mysql_connector_repackaged==0.3.1
pandas==2.2.3
Requests==2.32.3
selenium==4.29.0
SQLAlchemy==2.0.30
